refactor: props removed from chatinput
