+++ client/src/_context/room/room.type.ts
@@ -20,13 +20,11 @@ export interface IRoom {
     roomOutput: IOutput;
 }
 
-
 export interface IMessages {
     text: string;
     sender: string;
     notification?: boolean;
 }
-
 
 export interface State {
     _id: string;
@@ -37,7 +35,7 @@ export interface State {
     roomInput: string | null;
     roomLanguage: string;
     roomOutput: IOutput;
-    roomMessages : IMessages [];
+    roomMessages: IMessages[];
     roomLoaded: boolean | null;
     loading: boolean;
 
@@ -48,7 +46,7 @@ export interface State {
     updateRoomInput: (input: string) => void;
     updateRoomLanguage: (lang: string) => void;
     updateRoomOutput: (output: IOutput) => void;
-    updateMessages : (message : IMessages) => void
+    updateMessages: (message: IMessages) => void;
     setRoomUser: (username: string) => void;
     setLoading: (roomID: string) => void;
 }
