+++ client/src/components/Room/RoomDrawer/ChatInput.tsx
@@ -1,16 +1,38 @@
 import { Input } from 'antd';
+import socket from 'config/socket/socket';
+import { useSfx } from 'hooks';
 import React from 'react';
+import RoomContext from '_context/room/room.context';
 import style from './style.module.css';
 
 const { TextArea } = Input;
 
-type Props = {
-    sendMessage: () => void;
-    message: string;
-    onMessageChange: (event: React.FormEvent<HTMLTextAreaElement>) => void;
-};
+interface IMessage {
+    sender: string;
+    roomID: string;
+    text: string;
+}
+const ChatInput: React.FC = () => {
+    const { _id, currentUser } = React.useContext(RoomContext);
+    const [message, setMessage] = React.useState('');
+    const { playPop } = useSfx();
 
-const ChatInput: React.FC<Props> = ({ sendMessage, message, onMessageChange }) => {
+    const onMessageChange = (event: React.FormEvent<HTMLTextAreaElement>): void => {
+        setMessage(event.currentTarget.value);
+    };
+
+    const sendMessage = () => {
+        const body: IMessage = {
+            roomID: _id,
+            sender: currentUser,
+            text: message,
+        };
+        // send msg to server
+        socket.emit('realtime:message', body, () => setMessage(''));
+        playPop();
+    };
+
+
     return (
         <>
             <div className={style.chatInput}>
