+++ client/src/components/Room/RoomDrawer/Messages.tsx
@@ -3,27 +3,21 @@ import { Message } from 'layout';
 import React from 'react';
 import ScrollToBottom from 'react-scroll-to-bottom';
 import { v4 as uuidv4 } from 'uuid';
+import RoomContext from '_context/room/room.context';
 import style from './style.module.css';
 
-interface IMessages {
-    text: string;
-    sender: string;  
-    notification?: boolean;
-}
 
-type Props = {
-    messages: IMessages[];
-};
 
-const Messages: React.FC<Props> = ({ messages }) => {
-    return React.useMemo(() => {
+const Messages: React.FC = () => {
+    const {roomMessages} = React.useContext(RoomContext)
+
         return (
             <>
                 <Divider orientation='left' plain>
                     Messages
                 </Divider>
                 <ScrollToBottom className={style.messages}>
-                    {messages.map((message) => (
+                    {roomMessages.map((message) => (
                         <div key={uuidv4()}>
                             {message.notification && (
                                 <div className={style.message}>
@@ -31,7 +25,7 @@ const Messages: React.FC<Props> = ({ messages }) => {
                                     <Alert message={message.text} type='info' showIcon />
                                 </div>
                             )}
-    
+
                             {!message.notification && (
                                 <div className={style.message}>
                                     <Message text={message.text} sender={message.sender} />
@@ -42,8 +36,6 @@ const Messages: React.FC<Props> = ({ messages }) => {
                 </ScrollToBottom>
             </>
         );
-    }, [messages])
-   
 };
 
 export default Messages;
